Simplify exports for functions
I have simplified the exports for the functions in this Erlang
code. Originally, each export was on its own line, which
doesn't *really* make sense- especially when multiple exports can be
done at a time.

This solves the problem of code readability and shortening the LOC in
this Erlang source code, although its arguable that shortening the LOC
isn't a design goal.